+++ company-capf.el
@@ -25,10 +25,15 @@
 
 ;;; Code:
 
+(eval-when-compile (require 'cl))
+
 (defun company--capf-data ()
-  (let ((data (run-hook-wrapped 'completion-at-point-functions
-                                ;; Ignore misbehaving functions.
-                                #'completion--capf-wrapper 'optimist)))
+  ;; Ignore tags-completion-at-point-function because it subverts company-etags
+  ;; in the default value of company-backends, where the latter comes later.
+  (letf* (((default-value 'completion-at-point-functions) nil)
+          (data (run-hook-wrapped 'completion-at-point-functions
+                                  ;; Ignore misbehaving functions.
+                                  #'completion--capf-wrapper 'optimist)))
     (when (consp data) data)))
 
 (defun company-capf (command &optional arg &rest _args)